requirements.txt
# Dependencies
fastapi==0.104.1
pydantic[email]==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pyjwt==2.8.0
bcrypt==4.1.1
uvicorn==0.24.0
python-dotenv==1.0.0
